# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Believe Wireless API — powered by SignalWire
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface AvailablePhoneNumber {
  phoneNumber: string;
  friendlyName: string;
  region: string;
  rateCenter: string;
  monthlyFee: string;
}

export interface PhoneNumberClaim {
  phoneNumber: string;
  userEmail?: string;
  userName?: string;
}

export interface ClaimedPhoneNumber {
  sid: string;
  phoneNumber: string;
  friendlyName: string;
  status: string;
  claimedAt: string;
}

export interface Plan {
  id: string;
  name: string;
  price: string;
  description: string;
  features: string[];
  highlight: boolean;
  /** @nullable */
  badge?: string | null;
}

export type MessageDirection = typeof MessageDirection[keyof typeof MessageDirection];


export const MessageDirection = {
  inbound: 'inbound',
  'outbound-api': 'outbound-api',
} as const;

export interface Message {
  id: string;
  from: string;
  to: string;
  body: string;
  direction: MessageDirection;
  status: string;
  createdAt: string;
}

export interface MessageInput {
  from: string;
  to: string;
  body: string;
}

export interface WebhookPayload {
  From: string;
  To: string;
  Body: string;
  MessageSid?: string;
}

export interface Stats {
  totalNumbersClaimed: number;
  totalMessagesSent: number;
  totalUsers: number;
  coverageStates: number;
}

export interface AdminCredentials {
  email: string;
  password: string;
}

export interface AdminSession {
  email: string;
  loggedIn: boolean;
}

export interface AdminNumberItem {
  id: number;
  sid: string;
  phoneNumber: string;
  friendlyName: string;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  userName?: string | null;
  status: string;
  createdAt: string;
}

export interface AdminNumbersResponse {
  numbers: AdminNumberItem[];
  total: number;
}

export interface AdminMessageItem {
  id: number;
  sid: string;
  from: string;
  to: string;
  body: string;
  direction: string;
  status: string;
  createdAt: string;
}

export interface AdminMessagesResponse {
  messages: AdminMessageItem[];
  total: number;
}

export type ListAvailableNumbersParams = {
areaCode?: string;
contains?: string;
};

export type ListMessagesParams = {
phoneNumber: string;
limit?: number;
};

export type AdminListNumbersParams = {
limit?: number;
offset?: number;
};

export type AdminListMessagesParams = {
limit?: number;
offset?: number;
};