# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Believe Wireless API — powered by SignalWire
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/numbers:
    get:
      operationId: adminListNumbers
      tags: [admin]
      summary: Admin — list all claimed numbers
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer }
        - name: offset
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: All claimed numbers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminNumbersResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/messages:
    get:
      operationId: adminListMessages
      tags: [admin]
      summary: Admin — list all messages
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer }
        - name: offset
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: All messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMessagesResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    AdminCredentials:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
    AdminSession:
      type: object
      properties:
        email: { type: string }
        loggedIn: { type: boolean }
      required: [email, loggedIn]
    AdminNumberItem:
      type: object
      properties:
        id: { type: integer }
        sid: { type: string }
        phoneNumber: { type: string }
        friendlyName: { type: string }
        userEmail: { type: ["string", "null"] }
        userName: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
      required: [id, sid, phoneNumber, friendlyName, status, createdAt]
    AdminNumbersResponse:
      type: object
      properties:
        numbers:
          type: array
          items:
            $ref: "#/components/schemas/AdminNumberItem"
        total: { type: integer }
      required: [numbers, total]
    AdminMessageItem:
      type: object
      properties:
        id: { type: integer }
        sid: { type: string }
        from: { type: string }
        to: { type: string }
        body: { type: string }
        direction: { type: string }
        status: { type: string }
        createdAt: { type: string }
      required: [id, sid, from, to, body, direction, status, createdAt]
    AdminMessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AdminMessageItem"
        total: { type: integer }
      required: [messages, total]