# lib/api-zod/src/generated/types/adminMessageItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Believe Wireless API — powered by SignalWire
 * OpenAPI spec version: 0.1.0
 */

export interface AdminMessageItem {
  id: number;
  sid: string;
  from: string;
  to: string;
  body: string;
  direction: string;
  status: string;
  createdAt: string;
}