# lib/api-zod/src/generated/types/adminNumberItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Believe Wireless API — powered by SignalWire
 * OpenAPI spec version: 0.1.0
 */

export interface AdminNumberItem {
  id: number;
  sid: string;
  phoneNumber: string;
  friendlyName: string;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  userName?: string | null;
  status: string;
  createdAt: string;
}