# lib/api-zod/src/generated/types/adminSession.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Believe Wireless API — powered by SignalWire
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSession {
  email: string;
  loggedIn: boolean;
}